requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.3
boto3==1.37.14
pyarrow==19.0.1
pyspark==3.5.5
mlflow==2.21.0
scipy==1.15.2
pandas==2.2.3
PyYAML==6.0.2
prometheus-client==0.21.1
great-expectations==1.3.8
openlineage-python==1.28.0
jsonschema==4.23.0
pytest==8.3.5
pytest-mock==3.14.0
chispa==0.10.1
ruff==0.11.2
yamllint==1.35.1
